refactor: refactor lib/utils folder and add lib/mocks
