+++ src/components/ui/AnnouncementBar/announcementBar.spec.tsx
@@ -1,9 +1,8 @@
 import "intersection-observer";
-import { matchMedia } from "../../../lib/testUtils";
-import { renderWithContext, withPageBase } from "../../../lib/testUtils";
 import HomePage from "../../../pages/index";
 import { screen } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
+import { matchMedia, renderWithContext, withPageBase } from "@/lib/mocks";
 
 matchMedia();
