+++ src/lib/utils/debounce.ts
@@ -1,0 +1,22 @@
+// Debounce function
+// ref: https://github.com/you-dont-need/You-Dont-Need-Lodash-Underscore#_debounce
+
+const debounce = (func: any, wait: number, immediate?: boolean) => {
+  let timeout: NodeJS.Timeout;
+
+  return function () {
+    let context = this;
+    let args = arguments;
+
+    clearTimeout(timeout);
+
+    timeout = setTimeout(function () {
+      timeout = null;
+      if (!immediate) func.apply(context, args);
+    }, wait);
+
+    if (immediate && !timeout) func.apply(context, args);
+  };
+};
+
+export default debounce;
