+++ src/lib/utils/index.ts
@@ -1,0 +1,4 @@
+import handle from "./handle";
+import debounce from "./debounce";
+
+export { handle, debounce };
